.Removed some uneccessary files and views in the Folder. I have fixed some code to make it simpelr and easier. Fixed validation for Phone number in the Visitor Class by creating a phone number input mask in javascript.  I also made sure that the users had to create atleast one visitor inorder to book tickets or comment or else they wouldnt be able to do that and there would be a page telling them to create a visitor inorder to use those functions. Addiotnally, added the dropdown arrow hover to the VIsitor Log create and edit view and also styled up the VIsitor and VIsitorLog views. LAstly, have also added NoResult views to most of the tables just a few dont have them.

+++ ZooApp/Controllers/VisitorLogsController.cs
@@ -1,13 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ZooApp.Models;
-using ZooApp.data;
+using ZooApp.Data;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using ZooApp.data;
 
 namespace ZooApp.Controllers
 {
@@ -15,17 +16,29 @@ namespace ZooApp.Controllers
     public class VisitorLogsController : Controller
     {
         private readonly ZooAppContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public VisitorLogsController(ZooAppContext context)
+        public VisitorLogsController(ZooAppContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
-
         // GET: VisitorLogs
         public async Task<IActionResult> Index()
         {
-            var zooAppContext = _context.VisitorLogs.Include(v => v.Visitor);
-            return View(await zooAppContext.ToListAsync());
+            var userId = _userManager.GetUserId(User); // Get the current user's ID
+
+            var visitorLogs = await _context.VisitorLogs
+                .Include(vl => vl.Visitor)
+                .ToListAsync();
+
+            ViewData["VisitorCount"] = await _context.Visitor
+                .Where(v => v.CreatedByUserId == userId)
+                .CountAsync();
+
+            ViewBag.CurrentUserId = userId;
+
+            return View(visitorLogs); // Pass all visitorLogs to the view
         }
 
         // GET: VisitorLogs/Details/5
@@ -39,6 +52,7 @@ namespace ZooApp.Controllers
             var visitorLog = await _context.VisitorLogs
                 .Include(v => v.Visitor)
                 .FirstOrDefaultAsync(m => m.VisitorLogId == id);
+
             if (visitorLog == null)
             {
                 return NotFound();
@@ -50,7 +64,19 @@ namespace ZooApp.Controllers
         // GET: VisitorLogs/Create
         public IActionResult Create()
         {
-            ViewData["VisitorId"] = new SelectList(_context.Visitor, "VisitorId", "Name");
+            var userId = _userManager.GetUserId(User); // Get the current user's ID
+
+            var visitors = _context.Visitor
+                .Where(v => v.CreatedByUserId == userId)
+                .ToList();
+
+            if (!visitors.Any())
+            {
+                TempData["Message"] = "Please create a visitor first.";
+                return RedirectToAction("Index", "Visitors");
+            }
+
+            ViewData["VisitorId"] = new SelectList(visitors, "VisitorId", "Name");
             return View();
         }
 
@@ -62,11 +88,21 @@ namespace ZooApp.Controllers
             if (!ModelState.IsValid)
             {
                 visitorLog.CreatedOn = DateTime.Now;
+                visitorLog.VisitorId = _context.Visitor
+                    .Where(v => v.CreatedByUserId == _userManager.GetUserId(User))
+                    .Select(v => v.VisitorId)
+                    .FirstOrDefault();
+
                 _context.Add(visitorLog);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["VisitorId"] = new SelectList(_context.Visitor, "VisitorId", "Name", visitorLog.VisitorId);
+
+            var visitors = _context.Visitor
+                .Where(v => v.CreatedByUserId == _userManager.GetUserId(User))
+                .ToList();
+
+            ViewData["VisitorId"] = new SelectList(visitors, "VisitorId", "Name", visitorLog.VisitorId);
             return View(visitorLog);
         }
 
@@ -83,10 +119,18 @@ namespace ZooApp.Controllers
             {
                 return NotFound();
             }
-            ViewData["VisitorId"] = new SelectList(_context.Visitor, "VisitorId", "Name", visitorLog.VisitorId);
+
+            // Check if the current user is authorized to edit this log
+            var userId = _userManager.GetUserId(User); // Get the current user's ID
+
+
+            ViewData["VisitorId"] = new SelectList(_context.Visitor
+                .Where(v => v.CreatedByUserId == userId), "VisitorId", "Name", visitorLog.VisitorId);
+
             return View(visitorLog);
         }
 
+        // POST: VisitorLogs/Edit/5
         // POST: VisitorLogs/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -97,29 +141,51 @@ namespace ZooApp.Controllers
                 return NotFound();
             }
 
-            if (!ModelState.IsValid)
+            if (ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(visitorLog);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!VisitorLogExists(visitorLog.VisitorLogId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return View(visitorLog);
             }
-            ViewData["VisitorId"] = new SelectList(_context.Visitor, "VisitorId", "Name", visitorLog.VisitorId);
-            return View(visitorLog);
+
+            try
+            {
+                // Check if the current user is authorized to edit this log
+                var existingVisitorLog = await _context.VisitorLogs.FindAsync(id);
+                if (existingVisitorLog == null)
+                {
+                    return NotFound();
+                }
+
+                var userId = _userManager.GetUserId(User); // Get the current user's ID
+
+                // Ensure that only one instance of the entity is tracked by detaching the existing one
+                _context.Entry(existingVisitorLog).State = EntityState.Detached;
+
+                visitorLog.ModifiedOn = DateTime.Now;
+
+                // Update fields that can be modified
+                existingVisitorLog.VisitorId = visitorLog.VisitorId;
+                existingVisitorLog.Comments = visitorLog.Comments;
+                existingVisitorLog.Review = visitorLog.Review;
+                existingVisitorLog.ModifiedOn = visitorLog.ModifiedOn;
+
+                _context.Update(existingVisitorLog);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!VisitorLogExists(visitorLog.VisitorLogId))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            return RedirectToAction(nameof(Index));
         }
+
 
         // GET: VisitorLogs/Delete/5
         public async Task<IActionResult> Delete(int? id)
@@ -132,9 +198,17 @@ namespace ZooApp.Controllers
             var visitorLog = await _context.VisitorLogs
                 .Include(v => v.Visitor)
                 .FirstOrDefaultAsync(m => m.VisitorLogId == id);
+
             if (visitorLog == null)
             {
                 return NotFound();
+            }
+
+            // Check if the current user is authorized to delete this log
+            var userId = _userManager.GetUserId(User); // Get the current user's ID
+            if (visitorLog.Visitor.CreatedByUserId != userId)
+            {
+                return Forbid(); // Or handle unauthorized access as needed
             }
 
             return View(visitorLog);
@@ -148,6 +222,13 @@ namespace ZooApp.Controllers
             var visitorLog = await _context.VisitorLogs.FindAsync(id);
             if (visitorLog != null)
             {
+                // Check if the current user is authorized to delete this log
+                var userId = _userManager.GetUserId(User); // Get the current user's ID
+                if (visitorLog.Visitor.CreatedByUserId != userId)
+                {
+                    return Forbid(); // Or handle unauthorized access as needed
+                }
+
                 _context.VisitorLogs.Remove(visitorLog);
                 await _context.SaveChangesAsync();
             }
